## W03-P3: import sdata and students data from data_10.js

+++ demo/w01_dom_10/js/w01_10.js
@@ -13,7 +13,7 @@ const defaultResult = 0;
 let currentResult = defaultResult;
 
 function getUserInput() {
-  return parseInt(userInput.value);
+  return parseFloat(userInput.value);
 }
 
 function outputResult(result, text) {
